budget-backend/requirements.txt
flask
gunicorn
PyJWT
psycopg2-binary
passlib
flask-cors